requirements.txt
crewai
crewai-tools
tavily-python
streamlit
wikipedia
langchain_openai
langchain
